# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Inventory Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive]
        - name: lowStock
          in: query
          schema:
            type: boolean
        - name: sortBy
          in: query
          schema:
            type: string
            default: name
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedProducts"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update product
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductRequest"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Product deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        brand:
          type: ["string", "null"]
        hsnCode:
          type: ["string", "null"]
        unit:
          type: string
          enum: [gm, kg, ltr, ml, pcs, tonne, meter, feet, box, dozen, pack]
        purchasePrice:
          type: number
        sellingPrice:
          type: number
        mrp:
          type: number
        gstRate:
          type: number
        currentStock:
          type: number
        minStockAlert:
          type: number
        location:
          type: ["string", "null"]
        sku:
          type: ["string", "null"]
        supplierId:
          type: ["integer", "null"]
        supplierName:
          type: ["string", "null"]
        expiryDate:
          type: ["string", "null"]
          format: date
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - unit
        - purchasePrice
        - sellingPrice
        - mrp
        - gstRate
        - currentStock
        - minStockAlert
        - active
        - createdAt
    PaginatedProducts:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - data
        - total
        - page
        - limit
        - totalPages
    CreateProductRequest:
      type: object
      properties:
        name:
          type: string
        categoryId:
          type: integer
        brand:
          type: string
        hsnCode:
          type: string
        unit:
          type: string
          enum: [gm, kg, ltr, ml, pcs, tonne, meter, feet, box, dozen, pack]
        purchasePrice:
          type: number
        sellingPrice:
          type: number
        mrp:
          type: number
        gstRate:
          type: number
        currentStock:
          type: number
          default: 0
        minStockAlert:
          type: number
          default: 10
        location:
          type: string
        sku:
          type: string
        supplierId:
          type: integer
        expiryDate:
          type: string
          format: date
      required:
        - name
        - unit
        - purchasePrice
        - sellingPrice
        - mrp
        - gstRate
    UpdateProductRequest:
      type: object
      properties:
        name:
          type: string
        categoryId:
          type: integer
        brand:
          type: string
        hsnCode:
          type: string
        unit:
          type: string
          enum: [gm, kg, ltr, ml, pcs, tonne, meter, feet, box, dozen, pack]
        purchasePrice:
          type: number
        sellingPrice:
          type: number
        mrp:
          type: number
        gstRate:
          type: number
        currentStock:
          type: number
        minStockAlert:
          type: number
        location:
          type: string
        sku:
          type: string
        supplierId:
          type: integer
        expiryDate:
          type: string
          format: date
        active:
          type: boolean
...